refactor: taskDecorator 분리
