+++ src/main/java/com/sprint/mission/discodeit/common/config/async/SecurityContextTaskDecorator.java
@@ -1,29 +1,22 @@
-package com.sprint.mission.discodeit.common.config;
+package com.sprint.mission.discodeit.common.config.async;
 
-import java.util.Map;
 import org.slf4j.MDC;
 import org.springframework.core.task.TaskDecorator;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-public class ContextTaskDecorator implements TaskDecorator {
+public class SecurityContextTaskDecorator implements TaskDecorator {
 
   @Override
   public Runnable decorate(Runnable runnable) {
     SecurityContext securityContext = SecurityContextHolder.getContext();
-    Map<String, String> mdcContext = MDC.getCopyOfContextMap();
 
     return () -> {
       try {
         SecurityContextHolder.setContext(securityContext);
-        if (mdcContext != null) {
-          MDC.setContextMap(mdcContext);
-        }
-
         runnable.run();
       } finally {
         SecurityContextHolder.clearContext();
-        MDC.clear();
       }
     };
   }
