+++ src/main/java/com/sprint/mission/discodeit/domain/binarycontent/storage/local/LocalBinaryContentStorage.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.domain.binarycontent.storage.local;
 
-import com.sprint.mission.discodeit.common.util.FileUtils;
+import com.sprint.mission.discodeit.util.FileUtils;
 import com.sprint.mission.discodeit.domain.binarycontent.storage.BinaryContentStorage;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
