+++ src/main/java/com/sprint/mission/discodeit/domain/binarycontent/storage/s3/S3BinaryContentStorage.java
@@ -2,7 +2,7 @@ package com.sprint.mission.discodeit.domain.binarycontent.storage.s3;
 
 import static com.sprint.mission.discodeit.domain.binarycontent.entity.BinaryContentUploadStatus.*;
 
-import com.sprint.mission.discodeit.common.s3.S3Adapter;
+import com.sprint.mission.discodeit.s3.S3Adapter;
 import com.sprint.mission.discodeit.domain.binarycontent.entity.BinaryContent;
 import com.sprint.mission.discodeit.domain.binarycontent.exception.BinaryContentNotFoundException;
 import com.sprint.mission.discodeit.domain.binarycontent.repository.BinaryContentRepository;
