+++ src/main/java/com/sprint/mission/discodeit/s3/S3Adapter.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.common.s3;
+package com.sprint.mission.discodeit.s3;
 
 import java.io.InputStream;
 import java.net.URI;
