+++ src/main/java/com/sprint/mission/discodeit/s3/exception/S3CustomException.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.common.s3.exception;
+package com.sprint.mission.discodeit.s3.exception;
 
 import com.sprint.mission.discodeit.common.exception.DiscodeitException;
 import com.sprint.mission.discodeit.common.exception.ErrorCode;
