+++ src/test/java/com/sprint/mission/discodeit/domain/binarycontent/service/BinaryContentCoreTest.java
@@ -4,7 +4,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import com.sprint.mission.discodeit.common.s3.S3Adapter;
+import com.sprint.mission.discodeit.s3.S3Adapter;
 import com.sprint.mission.discodeit.domain.binarycontent.dto.BinaryContentRequest;
 import com.sprint.mission.discodeit.domain.binarycontent.entity.BinaryContent;
 import com.sprint.mission.discodeit.domain.binarycontent.exception.BinaryContentNotFoundException;
